Use u64 for coordinates to ensure deep zooms are represented

+++ fractals/fractal_utils/src/lib.rs
@@ -7,7 +7,7 @@ struct Color {
 }
 
 #[wasm_bindgen]
-pub fn mandelbrot(width: u32, height: u32, x_0: f32, x_1: f32, y_0: f32, y_1: f32, max_iterations: u8) -> Vec<u8> {
+pub fn mandelbrot(width: u32, height: u32, x_0: f64, x_1: f64, y_0: f64, y_1: f64, max_iterations: u8) -> Vec<u8> {
     println!("Computing fractal using width {} and height {}", width, height);
 
     //////// COLOR SCHEME ////////
@@ -64,12 +64,12 @@ pub fn mandelbrot(width: u32, height: u32, x_0: f32, x_1: f32, y_0: f32, y_1: f3
     let mut results_array = vec![0; size];
 
     //Convert the pixels into real coordinates in the Mandelbrot's range
-    let pixel_width  = (x_1 - x_0) / (width as f32);
-    let pixel_height = (y_1 - y_0) / (height as f32);
+    let pixel_width  = (x_1 - x_0) / (width as f64);
+    let pixel_height = (y_1 - y_0) / (height as f64);
     println!("Using pixel width {} and pixel height {}", pixel_width, pixel_height);
 
     //Formula values
-    let (mut x, mut y, mut x_squared, mut y_squared, mut iterations): (f32, f32, f32, f32, u8);
+    let (mut x, mut y, mut x_squared, mut y_squared, mut iterations): (f64, f64, f64, f64, u8);
     let mut pixel_x = x_0;
     let mut pixel_y = y_0;
     let mut pixel_idx: usize = 0; //Keeps track of our location in the array
